Documentation And Style Update
Updated the documentation to include a description for each documented
namespace in the README.  Added the format.txt formatting style
convention document to help others understand the convention in use.
Changed template parameter order for array memory allocation templates.
Unified the template parameter formatting in segmentation and
ordination.  Modified the text formatting of a for loop in trajection.